Improve performance of editor decorations
Instead of recreating the decoration & element every time, we cache the
results of creating the decoration and the underlying DOM node. Thus,
when the editor inserts it, it doesn't replace the node that was there
already and thus doesn't trigger the style recalculation that was
slowing down the editor while typing.